Style edits, pricing stability clarification, and extraneous info clean up
Ellipses in a paragraph is not a style we should adopt in official docs. 

Also, we make use of power provided by people other than "men". Remember to always use the full name of a thing before using the acronym on a single page. ASF info is extraneous and does not provide value (also technically incorrect to refer it to a grant, which has a technical definition, even if it semi-functions as a grant).

"A la carte" items are not within cloud.gov pricing at the moment, but the broader Infrastructure charge. Once cloud.gov's tooling actually contain these costs, we can list them here, but we also aren't explaining them yet, so listing them is confusing. Also, when we do, we should not use French to describe them. #politicalreality

Other important stylistic edits - the audience for this doc is already ingesting a tremendous amount of new information. Let's keep it to what they need to know, not our internal baseball.