Clarity on sandbox and cleanup edit
There was a question in doc about what the Sandbox is capped at. Calling it 1GB for now until someone tells me different. Also, it is *very* confusing  to ever say "unrecovered costs*. All costs are recovered. The Sandbox is recovered through upstream funding and sales.

Consistent term usage from the other pages, consistent capitalization, unnecessary formatting.

Skunkworks is never defined, and while a highly relevant concept to us, not a term for a general audience without clarification. "Prototyping" aligns with how that plan is usually requested, and how we talk about our work in that phase in other contexts.