important clarification about funding
Funds themselves are not severable or non-severable. Services are severable or non-severable, inherent to themselves. The important variable  to look at with federal funding is its "color", otherwise understood to be it's appropriation length (i.e. 1 year, 2 year, or no-year / indefinite). The color limits the time frame in which the money can be spent, regardless of the nature of the service or commodity. 

Since a non-severable and non-refundable services is actualized the moment a bona fide need is identified and the service is paid for, it does not expire at the end of the time frame of the underlying appropriation, because it was already _spent_ correctly according to its color.