Mild re-write
Edits for clarity / consistency / layman-ship (not sure if that's a word).

One thing I'm happy to go either way on is if we post links in the clear or hyperlink them to the plain text. I'm not sure we consistent on either style choice yet for our docs. For now, I'm a fan of hyperlinking and keeping full URLs out of the flow, but I'm not dogmatic if the consensus goes the other way.