Refer to a stable multi-buildpack fork which won't expire
The ddollar multi-buildpack is unmaintained, and has had a kill switch inserted so that any app using it, restaged or pushed after 1 Jan 2017, will break.

I've created a fork of this buildpack, specifically aimed at CF usage, which will be (a) stable and (b) actively developed.

This PR updates your docs to point towards the new fork. Obviously you don't have to do so, but there's a definite drop-dead date you should be aware of, regardless!