Fixing search by updating header and footer to be consistent
There were inconsistencies across the headers and footers on the site.
I have consolidated these into single items to make for easier
development.

This was some tech-debt that made sense to take on as the search was
only working on a few sections of the site as a result of these
separate header files.