move pricing terms to pricing
I set up an alias in case it’s linked anywhere else, but I also went
ahead and moved our own links to this page so they use the correct
directory. The directory it used to use is now empty. Didn’t want to
rely on that continuing to exist in any way.