reweight tracker to serve as index point
small thing, but weighting as zero makes it easy to determine “is this
overview content essential to read before reading about the fedramp
status or not?” if so, it gets a negative number. if not, it gets a
positive. because this is the most important thing for a lot of people,
and it’s definitely the biggest deal right now