changed name of files
This way we don’t have to change the links. We can update the
spreadsheet content later so it’s clear about date of update within the
content, but I’d like to get this committed for now, since that changes
pretty rarely and it’s already captured in metadata.