Try "cloud application platform"
Abby Kearns of the Cloud Foundry Foundation says that rather than trying to explain PaaS to people who might not know what IaaS and SaaS are, she concentrates on describing more what it *is* rather than what it's *not*. She said "Cloud Application Platform" seems to work well, and I agree that it focuses on the key attributes, particularly when paired with "for government teams". Let's try making this language tweak for a while and see if it helps.