elasticsearch56 plan name example command fix
The elasticsearch56 plan name appears to have changed from `medium` to `medium-ha`. It was changed in one place on the page, but not everywhere. I was following instructions from a project that was still using `medium` and was receiving this error `The plan medium could not be found for service elasticsearch56`. I came here to find the official plan names and notices the inconsistencies on the page.

I'm very new to cloud.gov, so please double check.