Update database backup and restoration guides
This changeset updates both the customer documentation and our operator runbook for database backups and restores with AWS RDS.  It clearly lays out how the backup and restoration process works, what information is needed, and how to perform the steps required to complete the work.

For customers, there is expanded information on how to connect directly to the service and perform manual backups for local restores.