Broker credential handling
During the work to approve the SCR for AWS Redis and Elasticsearch the
JAB TRs requested more information in the SSP and related documents on how we
satisfy IA-5 with broker credential management. This level of detail seemed
excessive for the SSP, so this is page is referenced in 10.1, under the
"Secrets" heading.