Adjust overflow scrolling of code blocks
This changeset updates our code blocks to allow for scrolling for content that expands beyond a visible line.  However, there are a couple of caveats:
- Modern browsers and OSes never display the scrollbars, even if you can scroll and set the overflow property to scroll
- It is not visibly clear that you can scroll

I am not sure if this is a better solution, but it is confusing when we have lines of code that get truncated and pushed to the next line, making it look like separate lines of code when they are not.  We may have to reformat some of our code blocks to account for this instead.