Update entity IDs to drop spurious protocol
Our entity ID does not actually include a protocol (https://). In a recent IDP integration, we discovered that including the protocol can cause an error due to the audience restriction not matching the actual audience (the entity ID). It appears some IDPs remove the protocol automatically and some do not. In any case, the correct value is the login hostname with no protocol.